+++ src/components/common/Pagination/Pagination.module.css
@@ -12,7 +12,7 @@
 
   .info {
     @media (max-width: 550px) {
-      margin-bottom: 8px;    
+      margin-bottom: 8px;
     }
   }
 }
